% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FishMaps2.R
\docType{data}
\name{dat.bum}
\alias{dat.bum}
\title{Blue marlin capture}
\format{A data frame with 496 rows and 4 variables:
\itemize{
  \item Quarter: quarter of the year
  \item Lat: latitude centered in the middle of 5x5 squares
  \item Lon: longitude centered in the middle of 5x5 squares
  \item BUM: blue marlin capture
}}
\source{
Based in ICCAT public databases
}
\usage{
data(dat.bum)
}
\description{
Blue marlin capture by the long-line fleet in Atlantic
Ocean
}
\keyword{datasets}

