% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMaps2.R
\docType{data}
\name{dat.lat.y}
\alias{dat.lat.y}
\title{Teste dataset}
\format{A data frame with 6557 rows and 5 variables:
\itemize{
  \item Year: year
  \item Quarter: quarter of the year
  \item Lat: latitude centered in the middle of 5x5 squares
  \item Lon: longitude centered in the middle of 5x5 squares
  \item Catch: some catch
}}
\source{
Based in ICCAT public databases
}
\usage{
data(dat.lat.y)
}
\description{
Teste dataset. PLEASE include more information.
}
\keyword{datasets}

