% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMaps2.R
\docType{data}
\name{dat.whm}
\alias{dat.whm}
\title{White marlin capture}
\format{A data frame with 496 rows and 4 variables:
\itemize{
  \item Quarter: quarter of the year
  \item Lat: latitude centered in the middle of 5x5 squares
  \item Lon: longitude centered in the middle of 5x5 squares
  \item WHM: white marlin capture
}}
\source{
Based in ICCAT public databases
}
\usage{
data(dat.whm)
}
\description{
White marlin capture by the long-line fleet in Atlantic
Ocean
}
\keyword{datasets}

