% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fishmap.R
\name{fishmap}
\alias{fishmap}
\title{Plots fishery data into maps}
\usage{
fishmap(x, y, fill, facetplot, dat, basemap, xlim, ylim, col.grd, bathymetry,
  ...)
}
\arguments{
\item{x}{A vector of coordinates (longitude)}

\item{y}{A vector of coordinates (latitude)}

\item{fill}{A numeric vector with data to fill in the map}

\item{facetplot}{wtf}

\item{dat}{The data frame containing the data}

\item{basemap}{The map datbase (from package \code{mapdata})}

\item{xlim,ylim}{X (longitude) and Y (latitude) limits of the map}

\item{col.grd}{The color of the grid}

\item{bathymetry}{A logical indicating if bathymetry lines should be
included}

\item{...}{Other arguments passed to \code{map}}
}
\value{
A map with data
}
\description{
Plots georeferenced fishery data (catch, effort, CPUE,
...) into maps. This description needs to be expanded.
}
\examples{
1 + 1
}

