% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMaps2.R
\docType{package}
\name{FishMaps2}
\alias{FishMaps2}
\alias{FishMaps2-package}
\title{Plots Fishery Data into Maps.}
\description{
Plots georeferenced fishery data (e.g. catch, effort
and CPUE) into maps. This is the lattice version of a previous
FishMaps version based on traditional grid graphics.
}

