% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilemap.R
\name{tilemap}
\alias{tilemap}
\title{Plots fishery data into maps}
\usage{
tilemap(x, y, z, data, facet.opt = NULL, xlim, ylim, col.fill = c("gray70",
  "gray10"), database = c("world", "worldHires"), bathymetry = FALSE, ...)
}
\arguments{
\item{x}{A vector of coordinates (longitude)}

\item{y}{A vector of coordinates (latitude)}

\item{z}{A numeric vector with data to fill in the map}

\item{data}{The data frame containing the data}

\item{facet.opt}{A list containing options to facet plots list(facet
= "~COLUMN NAME", ncol = number of columns to break plot)}

\item{xlim, ylim}{X (longitude) and Y (latitude) limits of the map}

\item{col.fill}{The color of the grid}

\item{database}{The map datbase (from package \code{mapdata})}

\item{bathymetry}{A logical indicating if bathymetry lines should be
included}

\item{...}{Other arguments passed to \code{map}}
}
\value{
A map with data
}
\description{
Plots georeferenced fishery data (catch, effort, CPUE,
...) into maps. This description needs to be expanded.
}
\examples{
1 + 1

}

