% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishMaps2.R
\docType{package}
\name{FishMaps2}
\alias{FishMaps2}
\alias{FishMaps2-package}
\title{Proportional symbol mapping for fishery data in batch mode}
\description{
Proportional symbol mapping for fishery data in batch mode (ggplot2 version)
}

